+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+## 3.2.0
+
 * Deprecating the `vat` parameter in `ItemDefinition` because of new [EU law for calculating VAT](http://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32006L0112&from=DE)
 * Deprecated `LaterPayClient.get_metered_access()` and `LaterPayClient.add_metered_access()` methods.
 * Added `return_url` / `failure_url` parameters for dialogs: https://www.laterpay.net/developers/docs/dialog-api